requirements.txt
Django==4.0.5
python-dotenv
djangorestframework==3.13.1
django-bootstrap3==21.2
djoser==2.0.5
djangorestframework-simplejwt==5.2.0
openai==0.27.0
psycopg2-binary
pytz